# OpenAPI/openapi.yaml
openapi: 3.0.1
info:
  title: MobileBff
  version: '1.0'
  description: BFF For Mobile App in Accounts and Payments
  contact:
    name: Godzilla Team
    url: http://localhost
    email: sarunas.kavaliauskas@seb.se
  x-seb:
    domain: DOM15
    tribe: TRI55
    art: ART86
    team: TEAM447
  x-seb-api:
    id: 84d513d3-3a76-4dd2-a78d-66eb108b11d7
  x-seb-portal:
    category: payments
    state: active
    summary: Provides BFF for SEB Mobile App
servers:
  - url: https://mobile-bff-ap-bff-mobile-apps.sebshift.dev.sebank.se
    x-environment: dev
  - url: https://mobile-bff-ap-bff-mobile-apps.sebshift.stg.sebank.se
    x-environment: stage
paths:
...
  '/corp/v1/accounts/{accountId}':
    get:
      tags:
        - CorporateAccounts
      parameters:
        - name: principal
          in: header
          schema:
            type: string
        - name: jwt-assertion
          in: header
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccount.CorporateGetAccountResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccount.CorporateGetAccountResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccount.CorporateGetAccountResponseModel'
        '206':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccount.CorporateGetAccountResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccount.CorporateGetAccountResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccount.CorporateGetAccountResponseModel'
  '/corp/v1/accounts/{accountId}/transactions':
    get:
      tags:
        - CorporateAccounts
      parameters:
        - name: principal
          in: header
          schema:
            type: string
        - name: jwt-assertion
          in: header
          schema:
            type: string
        - name: paginating_key
          in: query
          schema:
            type: string
        - name: paginating_size
          in: query
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccountTransactions.CorporateGetAccountTransactionsResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccountTransactions.CorporateGetAccountTransactionsResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccountTransactions.CorporateGetAccountTransactionsResponseModel'
        '206':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccountTransactions.CorporateGetAccountTransactionsResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccountTransactions.CorporateGetAccountTransactionsResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccountTransactions.CorporateGetAccountTransactionsResponseModel'
...
  '/private/v1/accounts/{accountId}':
    get:
      tags:
        - PrivateAccounts
      parameters:
        - name: principal
          in: header
          schema:
            type: string
        - name: jwt-assertion
          in: header
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccount.PrivateGetAccountResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccount.PrivateGetAccountResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccount.PrivateGetAccountResponseModel'
        '206':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccount.PrivateGetAccountResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccount.PrivateGetAccountResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccount.PrivateGetAccountResponseModel'
  '/private/v1/accounts/{accountId}/transactions':
    get:
      tags:
        - PrivateAccounts
      parameters:
        - name: principal
          in: header
          schema:
            type: string
        - name: jwt-assertion
          in: header
          schema:
            type: string
        - name: paginating_key
          in: query
          schema:
            type: string
        - name: paginating_size
          in: query
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccountTransactions.PrivateGetAccountTransactionsResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccountTransactions.PrivateGetAccountTransactionsResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccountTransactions.PrivateGetAccountTransactionsResponseModel'
        '206':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccountTransactions.PrivateGetAccountTransactionsResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccountTransactions.PrivateGetAccountTransactionsResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Private.GetAccountTransactions.PrivateGetAccountTransactionsResponseModel'
...
  '/youth/v1/accounts/{accountId}':
    get:
      tags:
        - YouthAccounts
      parameters:
        - name: jwt-assertion
          in: header
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccount.YouthGetAccountResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccount.YouthGetAccountResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccount.YouthGetAccountResponseModel'
        '206':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccount.YouthGetAccountResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccount.YouthGetAccountResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccount.YouthGetAccountResponseModel'
  '/youth/v1/accounts/{accountId}/transactions':
    get:
      tags:
        - YouthAccounts
      parameters:
        - name: jwt-assertion
          in: header
          schema:
            type: string
        - name: paginating_key
          in: query
          schema:
            type: string
        - name: paginating_size
          in: query
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccountTransactions.YouthGetAccountTransactionsResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccountTransactions.YouthGetAccountTransactionsResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccountTransactions.YouthGetAccountTransactionsResponseModel'
        '206':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccountTransactions.YouthGetAccountTransactionsResponseModel'
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccountTransactions.YouthGetAccountTransactionsResponseModel'
            text/json:
              schema:
                $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccountTransactions.YouthGetAccountTransactionsResponseModel'
...
components:
  schemas:
    MobileBff.Models.Corporate.GetAccount.CorporateDetailModel:
      type: object
      properties:
        title:
          type: string
          nullable: true
          readOnly: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccount.ItemModel'
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Corporate.GetAccount.CorporateGetAccountResponseModel:
      type: object
      properties:
        retrieved_date_time:
          type: string
          format: date-time
          readOnly: true
        resource_id:
          type: string
          nullable: true
          readOnly: true
        details:
          type: array
          items:
            $ref: '#/components/schemas/MobileBff.Models.Corporate.GetAccount.CorporateDetailModel'
          nullable: true
          readOnly: true
      additionalProperties: false
...
    MobileBff.Models.Corporate.GetAccountTransactions.CorporateGetAccountTransactionsResponseModel:
      type: object
      properties:
        retrieved_date_time:
          type: string
          format: date-time
          readOnly: true
        account:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.TransactionsAccountModel'
        paginating_information:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.PaginatingInformationModel'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.EntryModel'
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Private.GetAccount.PrivateDetailModel:
      type: object
      properties:
        title:
          type: string
          nullable: true
          readOnly: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccount.ItemModel'
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Private.GetAccount.PrivateGetAccountResponseModel:
      type: object
      properties:
        retrieved_date_time:
          type: string
          format: date-time
          readOnly: true
        resource_id:
          type: string
          nullable: true
          readOnly: true
        details:
          type: array
          items:
            $ref: '#/components/schemas/MobileBff.Models.Private.GetAccount.PrivateDetailModel'
          nullable: true
          readOnly: true
      additionalProperties: false
...
    MobileBff.Models.Private.GetAccountTransactions.PrivateEntryModel:
      type: object
      properties:
        transaction_id:
          type: string
          nullable: true
          readOnly: true
        value_date:
          type: string
          nullable: true
          readOnly: true
        booking_date:
          type: string
          nullable: true
          readOnly: true
        amount:
          type: number
          format: double
          nullable: true
          readOnly: true
        booked_balance:
          type: number
          format: double
          nullable: true
          readOnly: true
        currency:
          type: string
          nullable: true
          readOnly: true
        message:
          type: string
          nullable: true
          readOnly: true
        type:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.TypeModel'
        _links:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.LinksModel'
        external_id:
          type: string
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Private.GetAccountTransactions.PrivateGetAccountTransactionsResponseModel:
      type: object
      properties:
        retrieved_date_time:
          type: string
          format: date-time
          readOnly: true
        account:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.TransactionsAccountModel'
        paginating_information:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.PaginatingInformationModel'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/MobileBff.Models.Private.GetAccountTransactions.PrivateEntryModel'
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Shared.GetAccount.ItemModel:
      type: object
      properties:
        type:
          type: string
          nullable: true
          readOnly: true
        title:
          type: string
          nullable: true
        value:
          type: string
          nullable: true
        value_list:
          type: array
          items:
            type: string
          nullable: true
        action:
          type: string
          nullable: true
      additionalProperties: false
...
    MobileBff.Models.Shared.GetAccountTransactions.EntryModel:
      type: object
      properties:
        transaction_id:
          type: string
          nullable: true
          readOnly: true
        value_date:
          type: string
          nullable: true
          readOnly: true
        booking_date:
          type: string
          nullable: true
          readOnly: true
        amount:
          type: number
          format: double
          nullable: true
          readOnly: true
        booked_balance:
          type: number
          format: double
          nullable: true
          readOnly: true
        currency:
          type: string
          nullable: true
          readOnly: true
        message:
          type: string
          nullable: true
          readOnly: true
        type:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.TypeModel'
        _links:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.LinksModel'
      additionalProperties: false
    MobileBff.Models.Shared.GetAccountTransactions.LinksModel:
      type: object
      properties:
        transactionDetails:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.TransactionDetailsModel'
      additionalProperties: false
    MobileBff.Models.Shared.GetAccountTransactions.PaginatingInformationModel:
      type: object
      properties:
        paginating:
          type: boolean
          nullable: true
          readOnly: true
        date_from:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        paginating_key:
          type: string
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Shared.GetAccountTransactions.TransactionDetailsModel:
      type: object
      properties:
        href:
          type: string
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Shared.GetAccountTransactions.TransactionsAccountModel:
      type: object
      properties:
        resource_id:
          type: string
          nullable: true
          readOnly: true
        account_number:
          type: string
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Shared.GetAccountTransactions.TypeModel:
      type: object
      properties:
        code:
          type: string
          nullable: true
          readOnly: true
        text:
          type: string
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Youth.GetAccount.YouthDetailModel:
      type: object
      properties:
        title:
          type: string
          nullable: true
          readOnly: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccount.ItemModel'
          nullable: true
          readOnly: true
      additionalProperties: false
    MobileBff.Models.Youth.GetAccount.YouthGetAccountResponseModel:
      type: object
      properties:
        retrieved_date_time:
          type: string
          format: date-time
          readOnly: true
        resource_id:
          type: string
          nullable: true
          readOnly: true
        details:
          type: array
          items:
            $ref: '#/components/schemas/MobileBff.Models.Youth.GetAccount.YouthDetailModel'
          nullable: true
          readOnly: true
      additionalProperties: false
...
    MobileBff.Models.Youth.GetAccountTransactions.YouthGetAccountTransactionsResponseModel:
      type: object
      properties:
        retrieved_date_time:
          type: string
          format: date-time
          readOnly: true
        account:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.TransactionsAccountModel'
        paginating_information:
          $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.PaginatingInformationModel'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/MobileBff.Models.Shared.GetAccountTransactions.EntryModel'
          nullable: true
          readOnly: true
      additionalProperties: false